Merged `Title_Data` and `Title_Text`
Created a general system for applying text animations, and ported the sin
wave title to this system. This opens the door for any translation based
text animation. It doesn't support rotating the glyphs, but it could be
expanded to that in the future if it seems useful.

Also added the beginnings of a menu. Currently just displaying the text
with no interactions. I want to add keyboard and mouse support. The reason
for focusing on the animation system was to enable hover and select
animations.