Added keyboard navigation to the menu
Also Separated `.Select` into `.Key_Select` and `.Mouse_Select` to avoid
accidental clicks selecting random things. If the user was navigating
with the keyboard then clicking anywhere would select the hovered item.

Added a TODO to the animation_t update since it's speed seems to vary
based on the length of the text being animated. This is probably because
all the text shares the same timer, and it's updated by a fixed amount
per glyph.