Added delta time and freeing of the temp allocator
The temp allocator is freed at the end of each frame.

The delta is calculated at the end of each frame. This does mean that
delta will be zero for the first frame, but there isn't much you can do
about that. We need at least one frame to have passed before we can
calculate anything anyway.